Reorganized utils.php and created tec-utils.php that contains any code referencing tribe().
